Add files via upload
Questions:
Can you explain about DOM and the browser components?
Why we need an IP address and mac address?
Are JS and Java are the same?
What are the various Layout and JS engines?
What's a programming paradigm?
Features of JS?
What happens when you hit a website? Explain the flow in detail.
What are the HTTP error codes?
What is the diff between http1.1 vs http2?
What's an HTTP header?
What's the default browser method?